% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwrRasch.R
\docType{data}
\name{aid_st2}
\alias{aid_st2}
\title{Sample of test data from subtest 2 of the Adaptive Intelligence Diagnosticum (AID3; Kubinger \& Holocher-Ertl, 2014)}
\format{A data frame with 300 rows and 28 variables:
\itemize{
  \item ID: ID variable of each testee
  \item age_in_month: the age of the testperson in month
  \item sex: gender of the testee
  \item country: country of the testee
  \item stage: stage of the data collection
  \item it1...it18: items of the subtest 2
}}
\usage{
aid_st2
}
\description{
A dataset containing the test data of 300 childen (drawn randomly from the original dataset).
The variables are as follows:
}
\keyword{datasets}

