% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwrRasch.R
\docType{package}
\name{pwrRasch}
\alias{pwrRasch}
\alias{pwrRasch-package}
\title{Statistical Power Simulation for Testing the Rasch Model}
\description{
Statistical power simulation for testing the Rasch Model
based on a three-way analysis of variance design with mixed classification.
}
\author{
Takuya Yanagida [aut,cre] \email{takuya.yanagida@univie.ac.at},
 Jan Steinfeld [aut] \email{jan.steinfeld@univie.ac.at},
 Thomas Kiefer [ctb]

 Maintainer: Takuya Yanagida <takuya.yanagida@univie.ac.at>
}
\references{
Kubinger, K. D., Rasch, D., & Yanagida, T. (2009). On designing data-sampling for Rasch model calibrating an achievement test. \emph{Psychology Science Quarterly, 51}, 370-384.

Kubinger, K. D., Rasch, D., & Yanagida, T. (2011). A new approach for testing the Rasch model. \emph{Educational Research and Evaluation, 17}, 321-333.

Verhelst, N. D. (2008). An efficient MCMC algorithm to sample binary matrices with fixed marginals. \emph{Psychometrika, 73}(4), 705-728.

Verhelst, N., Hatzinger, R., & Mair, P. (2007). The Rasch sampler. \emph{Journal of Statistical Software, 20}(4), 1-14.
}
\seealso{
\link{aov.rasch}, \link{pwr.rasch}
}

